Search Flight Refactor
