+++ apps/openflight/backend/src/Controller/Flights/SearchFlightGetController.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
 namespace CodelyTv\Apps\OpenFlight\Backend\Controller\Flights;
 
 
-use CodelyTv\OpenFlight\Flights\Application\SearchFlightQuery;
+use CodelyTv\OpenFlight\Flights\Application\Search\SearchFlightQuery;
 use CodelyTv\Shared\Infrastructure\Symfony\ApiController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
