+++ src/OpenFlight/Flights/Application/Search/SearchFlightQuery.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\OpenFlight\Flights\Application;
+namespace CodelyTv\OpenFlight\Flights\Application\Search;
 
 use CodelyTv\Shared\Domain\Bus\Query\Query;
