+++ src/OpenFlight/Flights/Application/Search/SearchFlightQueryHandler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\OpenFlight\Flights\Application;
+namespace CodelyTv\OpenFlight\Flights\Application\Search;
 
 use CodelyTv\Shared\Domain\Bus\Query\QueryHandler;
 use CodelyTv\Shared\Domain\ValueObject\DateTimeValueObject;
