+++ src/OpenFlight/Flights/Application/Search/SearchFlightResponse.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace CodelyTv\OpenFlight\Flights\Application;
+namespace CodelyTv\OpenFlight\Flights\Application\Search;
 
 use CodelyTv\Shared\Domain\Bus\Query\Response;
